requirements.txt
streamlit
deepface
pillow
colorthief
scikit-learn
matplotlib
watchdog
numpy<2